pyproject.toml
[project]
name = "agentic-dungeon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.11.0",
    "networkx>=3.3",
    "matplotlib>=3.9.0",
    "tqdm>=4.67.1",
    "google-generativeai>=0.8.5",
    "tenacity>=9.1.2",
    "python-dotenv>=1.1.1",
    "fantasynamegenerator>=0.0.5",
    "fictional-names>=1.0.0",
    "loguru>=0.7.3",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.1",
]
